fix: resolve build errors after rebase on notification-commons refactor
