+++ packages/email-notification-dispatcher/src/handlers/eservices/handleEserviceArchivingScheduledToConsumer.ts
@@ -14,7 +14,7 @@ import {
   retrieveLatestDescriptor,
   retrieveTenant,
 } from "pagopa-interop-notification-commons";
-import { EServiceHandlerParams } from "../handlerCommons.js";
+import { EServiceHandlerParams } from "../../models/handlerParams.js";
 import { config } from "../../config/config.js";
 import { dateAtRomeZone } from "pagopa-interop-commons";
