+++ packages/email-notification-dispatcher/src/handlers/eservices/handleEserviceArchivingScheduledToProducer.ts
@@ -13,7 +13,7 @@ import {
   retrieveLatestDescriptor,
   retrieveTenant,
 } from "pagopa-interop-notification-commons";
-import { EServiceHandlerParams } from "../handlerCommons.js";
+import { EServiceHandlerParams } from "../../models/handlerParams.js";
 import { config } from "../../config/config.js";
 import { dateAtRomeZone } from "pagopa-interop-commons";
