+++ packages/email-notification-dispatcher/src/handlers/eservices/handleEserviceDescriptorArchivingCompletedToConsumer.ts
@@ -16,7 +16,7 @@ import {
   retrieveHTMLTemplate,
   retrieveTenant,
 } from "pagopa-interop-notification-commons";
-import { EServiceDescriptorHandlerParams } from "../handlerCommons.js";
+import { EServiceDescriptorHandlerParams } from "../../models/handlerParams.js";
 import { config } from "../../config/config.js";
 
 const notificationType: NotificationType = "eserviceStateChangedToConsumer";
