+++ packages/notification-commons/src/templates/email/templates.ts
@@ -91,6 +91,8 @@ export const eventMailTemplateType = {
     "producer-keychain-user-deleted-mail",
   producerKeychainNoKeysForAsyncEserviceMailTemplate:
     "producer-keychain-no-keys-for-async-eservice-mail",
+  asyncEserviceWithoutKeychainMailTemplate:
+    "async-eservice-without-keychain-mail",
   clientKeyAddedMailTemplate: "client-key-added-mail",
   producerKeychainKeyAddedMailTemplate: "producer-keychain-key-added-mail",
   newPurposeVersionQuotaAdjustmentRequestMailTemplate:
